backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-dotenv
python-multipart
passlib[bcrypt]
pyjwt
python-jose[cryptography]
psycopg2-binary
alembic